migrate diagnostics report
