+++ iOS/DuckDuckGo/RootDebugViewController.swift
@@ -261,6 +261,7 @@ class RootDebugViewController: UITableViewController {
     }
 }
 
+// Brindy - migrated to debug screen
 extension RootDebugViewController: DiagnosticReportDataSourceDelegate {
 
     func dataGatheringStarted() {
@@ -284,136 +285,6 @@ extension RootDebugViewController: DiagnosticReportDataSourceDelegate {
         DispatchQueue.main.async {
             self.reportGatheringActivity?.removeFromSuperview()
         }
-    }
-
-}
-
-protocol DiagnosticReportDataSourceDelegate: AnyObject {
-
-    func dataGatheringStarted()
-    func dataGatheringComplete()
-
-}
-
-class DiagnosticReportDataSource: UIActivityItemProvider {
-
-    weak var delegate: DiagnosticReportDataSourceDelegate?
-    var fireproofing: Fireproofing?
-
-    @UserDefaultsWrapper(key: .lastConfigurationRefreshDate, defaultValue: .distantPast)
-    private var lastRefreshDate: Date
-
-    convenience init(delegate: DiagnosticReportDataSourceDelegate, fireproofing: Fireproofing) {
-        self.init(placeholderItem: "")
-        self.delegate = delegate
-        self.fireproofing = fireproofing
-    }
-
-    override var item: Any {
-        delegate?.dataGatheringStarted()
-
-        let report = [reportHeader(),
-                      tabsReport(),
-                      imageCacheReport(),
-                      fireproofingReport(),
-                      configurationReport(),
-                      cookiesReport()].joined(separator: "\n\n")
-
-        delegate?.dataGatheringComplete()
-        return report
-    }
-
-    override func activityViewControllerLinkMetadata(_ activityViewController: UIActivityViewController) -> LPLinkMetadata? {
-        let metadata = LPLinkMetadata()
-        metadata.title = "Diagnostic Report"
-        return metadata
-    }
-
-    func reportHeader() -> String {
-        """
-        # DuckDuckGo App Diagnostic Report
-        Date: \(String(describing: Date()))
-        Version: \(AppVersion.shared.versionAndBuildNumber)
-        """
-    }
-
-    func fireproofingReport() -> String {
-        let allowedDomains = fireproofing?.allowedDomains.map { "* \($0)" } ?? []
-
-        let allowedDomainsEntry = ["### Allowed Domains"] + (allowedDomains.isEmpty ? [""] : allowedDomains)
-
-        return (["## Fireproofing Report"] + allowedDomainsEntry).joined(separator: "\n")
-    }
-
-    func imageCacheReport() -> String {
-        """
-        ## Image Cache Report
-        Bookmark Cache: \(Favicons.Constants.caches[.fireproof]?.count ?? -1)
-        Tabs Cache: \(Favicons.Constants.caches[.tabs]?.count ?? -1)
-        """
-    }
-
-    func configurationReport() -> String {
-        let etagStorage = DebugEtagStorage()
-        let configs = Configuration.allCases.map { $0.rawValue + ": " + (etagStorage.loadEtag(for: $0.storeKey) ?? "<none>") }
-        let lastRefreshDate = "Last refresh date: \(lastRefreshDate == .distantPast ? "Never" : String(describing: lastRefreshDate))"
-        return (["## Configuration Report"] + [lastRefreshDate] + configs).joined(separator: "\n")
-    }
-
-    func cookiesReport() -> String {
-        var cookies = [HTTPCookie]()
-
-        let group = DispatchGroup()
-        group.enter()
-        DispatchQueue.main.async {
-            WKWebsiteDataStore.current().httpCookieStore.getAllCookies { httpCookies in
-                cookies = httpCookies
-                group.leave()
-            }
-        }
-
-        var timeout = [String]()
-        if group.wait(timeout: .now() + 10) == .timedOut {
-            timeout = ["Failed to retrieve cookies in 10 seconds"]
-        }
-
-        let processedCookies = cookies
-            .sorted(by: { $0.domain < $1.domain })
-            .sorted(by: { $0.name < $1.name })
-            .map { $0.debugString }
-
-        return (["## Cookie Report"] + timeout + processedCookies).joined(separator: "\n")
-    }
-
-    func tabsReport() -> String {
-        """
-        ### Tabs Report
-        Tabs: \(TabsModel.get()?.count ?? -1)
-        """
-    }
-
-}
-
-private extension ImageCache {
-
-    var count: Int {
-        let url = diskStorage.directoryURL
-        do {
-            let contents = try FileManager.default.contentsOfDirectory(atPath: url.path)
-            return contents.count
-        } catch {
-            return -2
-        }
-    }
-
-}
-
-private extension HTTPCookie {
-
-    var debugString: String {
-        """
-        \(domain)\(path):\(name)=\(value.isEmpty ? "<blank>" : "<value>")\(expiresDate != nil ? ";expires=\(String(describing: expiresDate!))" : "")
-        """
     }
 
 }
